Move Persistence code into its own framework
The framework should be isolated from the rest of the project in a way
that if it has to be imported somewhere else, MacOS apps, another iOS
app or even iOS extensions, it can be done without dependencies. Note 
that the app that imports the dependencies is also responsible for
putting together the interfaces it's going to conform to.

This has also been updated to perform the writing and reading in a 
barrier async queue for better performance. It also now encodes and
decodes Codable protocols.

Some tests have been added to the Persistence class.