Create a shared lib to use with Siri Intents
Differently of the other libs in the project, this is just a target in
the App module. This is required since some files such as 
intents.intentdefinition need to be shared between all the targets 
associated to the application. It also makes sense in terms of 
architectural decision since all of this is part of the App layer.