Bind interfaces and use Repo in the extension
To ensure the same logic is happening in the app and its extension the
repository is being used instead of just hitting the API. Note that
it was necessary to bind some interfaces like it was done in the App
layer too. However the same binding file isn't being used because the
App layer should bind all the classes and the extension just the 
required ones.